+++ src/platform/platform_web.cpp
@@ -145,7 +145,7 @@ namespace Fusion
         glClear(GL_COLOR_BUFFER_BIT);
     }
 
-    float PlatformWeb::GetTime() const
+    double PlatformWeb::GetTime() const
     {
         return glfwGetTime();
     }
